+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fix the `deploy` CLI to the deploy new parameters with tags. [issue-18](https://github.com/lucasvieirasilva/aws-ssm-secrets-cli/issues/18)
 - Fix the `deploy` CLI to the secrets without `--confirm` option. [issue-17](https://github.com/lucasvieirasilva/aws-ssm-secrets-cli/issues/17)
+- Add `-d/--description` to the `set-parameter` and `set-secret` CLIs. [issue-16](https://github.com/lucasvieirasilva/aws-ssm-secrets-cli/issues/16)
 
 ## [1.0.1] - 2021-01-21